token and middelware

+++ laravel-api/routes/api.php
@@ -26,17 +26,21 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 //     'create', 'show', 'edit'
 // ]);
 
-Route::get('/products', [ProductsApiController::class, 'index']);
-Route::post('/products', [ProductsApiController::class, 'store']);
-Route::put('/products/{id}', [ProductsApiController::class, 'update']);
-Route::delete('/products/{id}', [ProductsApiController::class, 'destroy']);
-Route::get('/products/{id}', [ProductsApiController::class, 'show']);
+Route::get('/products', [ProductsApiController::class, 'index'])->middleware('auth:sanctum');
+Route::post('/products', [ProductsApiController::class, 'store'])->middleware('auth:sanctum');
+Route::put('/products/{id}', [ProductsApiController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/products/{id}', [ProductsApiController::class, 'destroy'])->middleware('auth:sanctum');
+Route::get('/products/{id}', [ProductsApiController::class, 'show'])->middleware('auth:sanctum');
 
 // Route::apiResource('products', ProductsApiController::class);
 
 
-Route::get('/users', [UserController::class, 'index']);
-Route::post('/users', [UserController::class, 'store']);
-Route::get('/users/{id}', [UserController::class, 'show']);
-Route::put('/users/{id}', [UserController::class, 'update']);
-Route::delete('/users/{id}', [UserController::class, 'destroy']);
+Route::get('/users', [UserController::class, 'index'])->middleware('auth:sanctum');
+// Route::post('/users', [UserController::class, 'store'])->middleware('auth:sanctum');
+Route::get('/users/{id}', [UserController::class, 'show'])->middleware('auth:sanctum');
+Route::put('/users/{id}', [UserController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/users/{id}', [UserController::class, 'destroy'])->middleware('auth:sanctum');
+
+
+Route::post('/auth/register', [UserController::class, 'createUser']);
+Route::post('/auth/login', [UserController::class, 'loginUser']);
